Loads in this commit. Bringing the repo up to current work.
I've sketched out the arbiter and made a small machine simulator.
I've also sketched out a multi-target device, not yet complete.
The state machine logic and beam parameter assertion element are very
rough but give an idea of where I am going. Missing is the element for
controlling the limit switches based on state and beam parameters.

I decided to create an arbiter that uses a hash table for components to
register their assertions. This allows for any component to ask for beam
parameters without having to hardwire in. The hashtables can also be
viewed by EPICS diagnostics to see who is asserting beam parameters.

At the time of this commit I was in the middle of creating a globally
accessible assertion pool (ie. the hashtable) and using its
methods/actions within the beam class transition manager to add and
remove assertions accordingly. Once this is hooked up sensibly I should
be able to change the device state and see the machine simulator respond
by adjusting attenuation, etc.